specify correct folder in circle yml, move tests into tests folder
