+++ circle.yml
@@ -10,7 +10,7 @@ test:
   override:
     - npm run coverage:ci
   post:
-    - mv ./actual $CIRCLE_ARTIFACTS
+    - mv ./test/actual $CIRCLE_ARTIFACTS
 database:
   override:
     - mysql -u ubuntu -e "create database test"
