+++ test/tests/integration.test.ts
@@ -1,7 +1,7 @@
-/// <reference path="../node_modules/@types/mocha/index.d.ts" />
+/// <reference path="../../node_modules/@types/mocha/index.d.ts" />
 import * as assert from 'power-assert'
-import { Database, getDatabase } from '../src/index'
-import { writeTsFile, compare, loadSchema } from './testUtility'
+import { Database, getDatabase } from '../../src/index'
+import { writeTsFile, compare, loadSchema } from '../testUtility'
 
 describe('schemat generation integration testing', () => {
     describe('postgres', () => {
@@ -11,7 +11,7 @@ describe('schemat generation integration testing', () => {
                 return this.skip()
             }
             db = getDatabase(process.env.POSTGRES_URL)
-            await loadSchema(db, `${__dirname}/fixture/postgres/initCleanup.sql`)
+            await loadSchema(db, './test/fixture/postgres/initCleanup.sql')
         })
 
         it('Basic generation', async () => {
@@ -39,7 +39,7 @@ describe('schemat generation integration testing', () => {
                 return this.skip()
             }
             db = getDatabase(`${process.env.MYSQL_URL}?multipleStatements=true`)
-            await loadSchema(db, `${__dirname}/fixture/mysql/initCleanup.sql`)
+            await loadSchema(db, './test/fixture/mysql/initCleanup.sql')
         })
         it ('Basic generation', async () => {
             const inputSQLFile = 'test/fixture/mysql/osm.sql'
