+++ test/tests/unit.test.ts
@@ -1,5 +1,5 @@
 import * as ts from 'typescript';
-import { compile } from './testUtility'
+import { compile } from '../testUtility'
 
 describe('end user use case', () => {
     it('usecase.ts should compile without error', () => {
